Clothing Size Calculator
This project is a group project. Our group was tasked to create a program to solve a real-world problem at random and since we were a group of ladies, we singled out the challenges we face. One of them was finding the right dress size and a pair of trousers for our individual body shapes.
In conclusion, we created the blueprint for an app which recommends the dress size and pants size based on the measurements you provide in inches.
Hope others can relate!